requirements.txt
# Data Extraction & API
spotipy==2.23.0
requests==2.31.0

# Database
duckdb==0.10.0

# Data Transformation
dbt-core==1.7.9
dbt-duckdb==1.7.2

# Dashboard & Visualization
streamlit==1.31.1
plotly==5.19.0
pandas==2.2.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1

# Development & Testing
pytest==8.0.1
black==24.2.0
